Added Cash Payment clause to transaction code algorithm
The code in its current state will not execute the cash payment, but will return to the front end the total amount of cash required for the user to pay by adding all the cash payments in all the accounting entries. The front end person is then meant to trigger a handshake with a payment platform such as PayStack to request that the user pays this amount. I will create a new method for the front end to call once it can send the transactionID of the succesfully completed payment. This new method will then continue the warrant based accounting entries while having knowledge of the completion of the cash payment. However, if it was the MOE meant to pay cash, and not receive, this method in its current state will simply return "MOE cannot make cash payments until we integrate with a bank"